(feat): support i8 and i16

+++ crates/parser/src/logic/arithmetic.rs
@@ -11,7 +11,7 @@
 //!
 //! let expr = Arithmetic::new(
 //!     Factor::Var("x".into()),
-//!     vec![(ArithmeticOperator::Plus, Factor::Const(ConstType::Int32(5)))],
+//!     vec![(ArithmeticOperator::Plus, Factor::Const(ConstType::Int(5)))],
 //! );
 //! assert_eq!(expr.to_string(), "x + 5");
 //! ```
@@ -225,15 +225,15 @@ impl Lexeme for Arithmetic {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::primitive::ConstType::{Int32, Text};
+    use crate::primitive::ConstType::{Int, Text};
     use ArithmeticOperator::*;
     use Factor::*;
 
     fn v(name: &str) -> Factor {
         Var(name.into())
     }
-    fn i(n: i32) -> Factor {
-        Const(Int32(n))
+    fn i(n: i64) -> Factor {
+        Const(Int(n))
     }
     fn s(t: &str) -> Factor {
         Const(Text(t.into()))
